Fix reading footer of ORC file when stripe has more than Integer.MAX_VALUE rows

+++ lib/trino-orc/src/main/java/io/trino/orc/OrcRecordReader.java
@@ -748,7 +748,7 @@ public class OrcRecordReader
         }
     }
 
-    private void validateWriteStripe(int rowCount)
+    private void validateWriteStripe(long rowCount)
     {
         writeChecksumBuilder.ifPresent(builder -> builder.addStripe(rowCount));
     }
